+++ NEWS.md
@@ -37,5 +37,6 @@
 - Cross with ASD now tracks body temporarily.
 - Change order of instantiating Cross Widget crosser attribute.
 - Not sending logger to ConnectASDWorker Tracker, as it would close them when stopping the tracking.
+- ConnectASDWorker Tracker has logger that logs only to the logfile.
 
 ## News